Extract StatusPending and StatusEnded components from Status
